Implement financial news aggregator for watchlist
This script aggregates financial news from Yahoo Finance API for a predefined watchlist of companies, fetching and cleaning news data, and sorting it chronologically for display. 
This is my initial version where I struggled with exception handling when dealing with messy data, and especially the time of the news, which I ended up using 3 different methods of finding the timestamps in order to make it work.